Refactoring IGDB API
