+++ app/src/main/java/com/insa/mygamelist/data/GameViewModel.kt
@@ -1,0 +1,21 @@
+package com.insa.mygamelist.data
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.launch
+
+class GameViewModel : ViewModel() {
+    private val repository = IGDBServiceAPI()
+
+    private val _games = MutableLiveData<List<GameAPI>>()
+    val games: LiveData<List<GameAPI>> get() = _games
+
+    fun fetchGames() {
+        viewModelScope.launch {
+            val gamesList = repository.getGames()
+            _games.postValue(gamesList ?: emptyList())
+        }
+    }
+}
